# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Freshify API Documentation
  description: |-
    An API for creating and managing a household inventory and shopping list.
    The API is written in Java using the Spring Boot framework. The API uses a
    MySQL database. The API is documented using Swagger. The API is hosted using
    Apache.
    Some useful links:
    - [The GitLab repository](https://gitlab.stud.idi.ntnu.no/idatt2106-v23-10/smartmat-backend) 
  contact:
    email: vkbugge@hotmail.com
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /user/{id}:
    get:
      tags:
        - account
      summary: Get user by id
      description: Returns a single user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: ID of user to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFull'
        '404':
          description: User not found
    delete:
      tags:
        - account
      summary: Delete user
      description: Deletes a user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID of user to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: successful operation
        '401':
          description: Access denied
        '404':
          description: User not found
...
  /household:
    post:
      tags:
        - household
      summary: Create a household
      description: Creates a household
      operationId: createHousehold
      requestBody:
        description: The household to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Household'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '409':
          description: Household already exists
...
  /household/{id}/inventory:
    post:
      tags:
        - inventory
      summary: Add item to inventory
      description: Takes a list of entries, each entry contains an item type id and count. This way multiple items can be added.
      operationId: addInventoryItem
      parameters:
        - name: id
          in: path
          description: ID of the household to add items to.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: List of items to be added to inventory
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  itemTypeId:
                    type: integer
                    format: int64
                  count:
                    type: integer
                    format: int64
      responses:
        '201':
          description: Returns a list of the newly added items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItems'
        '401':
          description: Unauthorized response, user is not validated or does not have access
        '404':
          description: Household not found/User not found/Item type not found
    get:
      tags:
       - inventory
      summary: Gets all items from inventory
      description: Gets all items of a specific inventory
      operationId: getInventoryItems
      parameters:
        - name: id
          in: path
          description: ID of household's shopping list to get items from
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: item found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          description: Unauthorized
        '404':
          description: Household not found
    put:
      tags:
        - inventory
      summary: Alters an item in inventory
      description: Alters an item in inventory. Either the state or the remaining amount can be changed.
      operationId: updateInventoryItem
      parameters:
        - name: id
          in: path
          description: ID of household's inventory to update the item in
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The item to be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItem"
      responses:
        '200':
          description: Successful operation
        '400':
          description: The given item status is not a valid status for item/item does not belong to household
        '401':
          description: Unauthorized response, user is not validated or does not have access
        '404':
          description: Item not found/household not found
  /household/{id}/inventory/{itemId}:
    delete:
      tags:
        - inventory
      summary: Delete item from inventory
      description: Delete an item from inventory of a household.
      operationId: deleteInventoryItem
      parameters:
        - name: id
          in: path
          description: ID of household's inventory to delete the item from 
          required: true
          schema:
            type: integer
            format: int64
        - name: itemId
          in: path
          description: ID of item to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized response, user is not validated or does not have access
        '404':
          description: Item not found/household not found
  /household/{id}/inventory/suggest:
    post:
      tags:
        - inventory
      summary: Add suggestion to inventory
      description: Takes a list of entries, each entry contains an item type id and count. These item types are added as suggestions to the inventory.
      operationId: addInventorySuggestion
      parameters:
        - name: id
          in: path
          description: ID of the household to add items to.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: List of suggestions to be added to inventory
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  itemTypeId:
                    type: integer
                    format: int64
                  count:
                    type: integer
                    format: int64
      responses:
        '201':
          description: Returns a list of the newly added suggestions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItems'
        '401':
          description: Unauthorized response, user is not validated or does not have access
        '404':
          description: Household not found/User not found/Item type not found
...
components:
  schemas:
...
    UserFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        email:
          type: string
          format: email
          example: example@example.org
        firstName:
          type: string
          example: John
    ItemType:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "Milk"
...
    ItemState:
      type: string
      enum:
        - INVENTORY
        - USED
    UpdateItem:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        remaining:
          description: The proportion of the item that was wasted
          type: number
          format: double
          example: 0.5
        state:
          description: The state of the item
          $ref: "#/components/schemas/ItemState"
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 123
        state:
          $ref: '#/components/schemas/ItemState' 
        remaining:
          type: number
          format: double
          example: 0.5
        bought:
          type: string
          format: date
          example: "2020-01-01"
        suggested:
          type: boolean
        type:
          $ref: '#/components/schemas/ItemType'
        addedBy:
          $ref: '#/components/schemas/UserFull'
    InventoryItems:
      type: array
      items:
          properties:
            id:
              type: integer
              example: 123
            state:
              $ref: '#/components/schemas/ItemState' 
            remaining:
              type: number
              format: double
              example: 1
            bought:
              type: string
              format: date
              example: "2020-01-01"
            suggested:
              type: boolean
            type:
              $ref: '#/components/schemas/ItemType'
            addedBy:
              $ref: '#/components/schemas/UserFull'
...
    Household:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "My family"